Implemented recovery mechanism for  GuessWord
It was required to implement a functionality that allows the user to
enter data (by using the data entry functionality) if the data file doesn't exist / is empty / contains invalid
data at application startup

If the file doesn't exist it will be created as empty file. Then the
user will be able to access the data entry menu automatically to enter
new words. Then these words would get loaded and guessing will commence

Also it was required to do some updates of the output/error messages and
some input/output/error functionality refactoring